LayoutTests/fast/regions/offsetLeft-offsetTop-in-region-float-expected.txt

Test offsetLeft and offsetTop in a named flow with floats.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS divText1.offsetParent is document.body
PASS divText1.offsetLeft is 200
PASS divText1.offsetTop is 300
PASS divText2.offsetParent is document.body
PASS divText2.offsetLeft is 300
PASS divText2.offsetTop is 500
PASS successfullyParsed is true

TEST COMPLETE
